pyproject.toml

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "veeam-cli"
version = "0.1.0"
description = "A CLI for interacting with the Veeam Backup & Replication API."
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "click",
    "requests",
    "python-dotenv",
    "rich",
]

[project.scripts]
veeam-cli = "veeam_cli.cli:cli"
